Refactor -  Make code more accessible and modular
