+++ client/constants/apiConstants.js
@@ -1,0 +1,1 @@
+export const apiKey = '49a19ed77c416266d2370caeb7f280d9.33105';
