+++ client/constants/graphConstants.js
@@ -1,0 +1,3 @@
+export const denverLatitude = 39.7392;
+export const denverLongitude = -104.9903;
+export const defaultZoom = 10;
